+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2022-05-25
+### Added
+- Support for custom outlet names
+- Button to plot all PVs
+
+### Changed
+- Opens power data in new tab
+
 ## [0.4.0] - 2022-02-11
 ### Added
 - User can now unlink devices or their Telegram ID from notifications